requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.1.7
django-model-utils==4.3.1
django-notifications-hq==1.7.0
django-resized==1.0.2
easy-thumbnails==2.8.5
idna==3.4
jsonfield==3.1.0
paynow==1.0.4
Pillow==9.4.0
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
swapper==1.3.0
urllib3==1.26.14
